Rename ExtractFrames to ExtractWindows
