--- dali/kernels/signal/window/extract_frames_cpu.h
@@ -1,90 +1,0 @@
-// Copyright (c) 2019, NVIDIA CORPORATION. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef DALI_KERNELS_SIGNAL_WINDOW_EXTRACT_FRAMES_H_
-#define DALI_KERNELS_SIGNAL_WINDOW_EXTRACT_FRAMES_H_
-
-#include <memory>
-#include <complex>
-#include "dali/core/common.h"
-#include "dali/core/error_handling.h"
-#include "dali/core/format.h"
-#include "dali/core/util.h"
-#include "dali/kernels/kernel.h"
-
-namespace dali {
-namespace kernels {
-namespace signal {
-namespace window {
-
-struct ExtractFramesArgs {
-  int64_t window_length = -1;
-  int64_t window_step = -1;
-  int64_t in_time_axis = -1;
-  int64_t out_frame_axis = -1;
-
-  inline bool operator==(const ExtractFramesArgs& oth) const {
-    return window_length == oth.window_length &&
-           window_step == oth.window_step &&
-           in_time_axis == oth.in_time_axis &&
-           out_frame_axis == oth.out_frame_axis;
-  }
-
-  inline bool operator!=(const ExtractFramesArgs& oth) const {
-    return !operator==(oth);
-  }
-};
-
-/**
- * @brief Extracts windows from 1D signal applying a custom window function
- *
- * @param args.window_length Window size in number of samples
- *
- * @param args.window_step Length of the step between windows. If not provided, win_length
- *        will be choosen (no overlap)
- */
-template <typename OutputType = float, typename InputType = float, int Dims = 1>
-class DLL_PUBLIC ExtractFramesCpu {
- public:
-  static constexpr int InputDims = Dims;
-  static constexpr int OutputDims = Dims + 1;
-
-  static_assert(std::is_same<OutputType, InputType>::value,
-    "Type conversion is not allowed in this kernel");
-
-  DLL_PUBLIC ~ExtractFramesCpu();
-
-  DLL_PUBLIC KernelRequirements Setup(KernelContext &context,
-                                      const InTensorCPU<InputType, InputDims> &in,
-                                      const ExtractFramesArgs &args);
-
-  DLL_PUBLIC void Run(KernelContext &context,
-                      const OutTensorCPU<OutputType, OutputDims> &out,
-                      const InTensorCPU<InputType, InputDims> &in,
-                      const ExtractFramesArgs &args);
- private:
-  ExtractFramesArgs args_;
-  int64_t window_length_ = -1;
-  int64_t window_step_ = -1;
-  int64_t nwindows_ = -1;
-  int64_t in_time_axis_ = -1;
-  int64_t out_frame_axis_ = -1;
-};
-
-}  // namespace window
-}  // namespace signal
-}  // namespace kernels
-}  // namespace dali
-
-#endif  // DALI_KERNELS_SIGNAL_WINDOW_EXTRACT_FRAMES_H_
