+++ dali/kernels/signal/window/extract_windows_cpu.cc
@@ -14,13 +14,12 @@
 
 #include "dali/kernels/signal/window/extract_windows_cpu.h"
 #include <algorithm>
+#include "dali/kernels/signal/signal_kernel_utils.h"
 #include "dali/core/common.h"
 #include "dali/core/error_handling.h"
 #include "dali/core/format.h"
 #include "dali/core/util.h"
 #include "dali/kernels/kernel.h"
-#include "dali/kernels/common/for_axis.h"
-#include "dali/kernels/common/utils.h"
 
 namespace dali {
 namespace kernels {
@@ -40,45 +39,48 @@ template <typename OutputType, typename InputType, int Dims>
 KernelRequirements ExtractWindowsCpu<OutputType, InputType, Dims>::Setup(
     KernelContext &context,
     const InTensorCPU<InputType, InputDims> &in,
-    const InTensorCPU<float, 1> &window_fn,
     const ExtractWindowsArgs &args) {
   KernelRequirements req;
 
   window_length_ = args.window_length > 0 ? args.window_length : 1;
   window_step_ = args.window_step > 0 ? args.window_step : 1;
-  window_center_offset_ = args.window_center < 0 ? window_length_ / 2 : args.window_center;
-  reflect_pad_ = args.reflect_pad;
-
-  DALI_ENFORCE(window_center_offset_ >= 0 && window_center_offset_ <= window_length_,
-    make_string("Window center offset must be in the range [0, ", window_length_, "]"));
-
-  window_fn_length_ = volume(window_fn.shape);
-  DALI_ENFORCE(window_fn_length_ > 0, "Window function should not be empty");
-  DALI_ENFORCE(window_fn_length_ <= window_length_,
-    "Window function size should be equal or less than the specified window length");
 
   // input data temporal axis (last in input shape by default)
-  axis_ = args.axis >= 0 ? args.axis : InputDims - 1;
-  DALI_ENFORCE(axis_ >= 0 && axis_ < InputDims,
-    make_string("Input temporal axis (", axis_, ") is out of range [0, ", InputDims, ")"));
-  DALI_ENFORCE(axis_ == InputDims - 1,
+  in_time_axis_ = args.in_time_axis >= 0 ? args.in_time_axis : InputDims - 1;
+  DALI_ENFORCE(in_time_axis_ >= 0 && in_time_axis_ < InputDims,
+    make_string("Input temporal axis (", in_time_axis_, ") is out of range [0, ", InputDims, ")"));
+  DALI_ENFORCE(in_time_axis_ == InputDims - 1,
     "Current implementation expects time dimension to be the inner-most dimension");
 
-  const auto n = in.shape[axis_];
+  // frame temporal axis (last in output shape by default)
+  out_frame_axis_ = args.out_frame_axis >= 0 ? args.out_frame_axis : OutputDims - 1;
+  DALI_ENFORCE(out_frame_axis_ >= 0 && out_frame_axis_ < OutputDims,
+    make_string("Output frame temporal axis (", out_frame_axis_, ") is out of range [0, ",
+      OutputDims, ")"));
+  DALI_ENFORCE(out_frame_axis_ == OutputDims - 1,
+    "Current implementation expects window time dimension to be the inner-most dimension");
 
-  nwindows_ =  n / window_step_ + 1;
+
+  const auto n = in.shape[in_time_axis_];
+
+  nwindows_ = (n + window_step_ - 1) / window_step_;
   assert(nwindows_ > 0);
 
   TensorShape<DynamicDimensions> out_shape;
   out_shape.resize(OutputDims);
 
   for (int d = 0, out_idx = 0, in_idx = 0; out_idx < OutputDims; d++) {
-    if (d == axis_) {
-      assert(out_idx + 1 < OutputDims);
-      assert(in_idx < InputDims);
-      out_shape[out_idx++] = window_length_;
-      out_shape[out_idx++] = nwindows_;
-      in_idx++;
+    if (d == out_frame_axis_ || d == in_time_axis_) {
+      if (d == out_frame_axis_) {
+        assert(out_idx < OutputDims);
+        out_shape[out_idx++] = window_length_;
+      }
+      if (d == in_time_axis_) {
+        assert(out_idx < OutputDims);
+        assert(in_idx < InputDims);
+        out_shape[out_idx++] = nwindows_;
+        in_idx++;
+      }
     } else {
       assert(out_idx < OutputDims);
       assert(in_idx < InputDims);
@@ -94,7 +96,6 @@ void ExtractWindowsCpu<OutputType, InputType, Dims>::Run(
     KernelContext &context,
     const OutTensorCPU<OutputType, OutputDims> &out,
     const InTensorCPU<InputType, InputDims> &in,
-    const InTensorCPU<float, 1> &window_fn,
     const ExtractWindowsArgs &args) {
 
   auto in_shape = in.shape;
@@ -103,42 +104,34 @@ void ExtractWindowsCpu<OutputType, InputType, Dims>::Run(
   // flat_out_shape is the output shape with both window index and time dimensions combined into
   // one dimension
   auto flat_out_shape = in_shape;
-  flat_out_shape[axis_] = nwindows_ * window_length_;
+  flat_out_shape[in_time_axis_] = nwindows_ * window_length_;
   auto out_strides = GetStrides(flat_out_shape);
+
+  std::vector<std::pair<OutputType*, const InputType*>> slices;
+  slices.push_back(std::make_pair(out.data, in.data));
+  Get1DSlices(slices,
+              flat_out_shape.data(), out_strides.data(),
+              in_shape.data(), in_strides.data(),
+              in_time_axis_, InputDims);
+
+  const auto out_stride = out_strides[in_time_axis_];
+  const auto in_stride = in_strides[in_time_axis_];
+  const auto in_t_size = in.shape[in_time_axis_];
 
   ForAxis(
     out.data, in.data, flat_out_shape.data(), out_strides.data(),
-    in_shape.data(), in_strides.data(), axis_, InputDims,
-    [this, &window_fn](
+    in_shape.data(), in_strides.data(), in_time_axis_, InputDims,
+    [this](
       OutputType *out_data, const InputType *in_data,
       int64_t out_size, int64_t out_stride, int64_t in_size, int64_t in_stride) {
-        for (int64_t window_idx = 0; window_idx < nwindows_; window_idx++) {
-          int64_t window_start = window_idx * window_step_ - window_center_offset_;
-          // Window needs special treatment (falls outside of the signal)
-          if (window_start < 0 || window_start + window_length_ > in_size) {
-            for (int64_t t = 0; t < window_length_; t++) {
-              int64_t out_idx = t * nwindows_ + window_idx;
-              int64_t in_idx = window_start + t;
-              if (reflect_pad_) {
-                // find the mirrored position if the index is out of bounds
-                while (in_idx < 0 || in_idx >= in_size) {
-                  if (in_idx < 0) {
-                    // left side
-                    in_idx = -in_idx;
-                  } else {
-                    // right side
-                    in_idx = 2 * in_size - 2 - in_idx;
-                  }
-                }
-              }
-              out_data[out_idx * out_stride] = (in_idx >= 0 && in_idx < in_size) ?
-                window_fn.data[t] * in_data[in_idx * in_stride] : 0;
-            }
-          } else {  // no special treatment for this window (just copy)
-            for (int64_t t = 0; t < window_length_; t++) {
-              int64_t out_idx = t * nwindows_ + window_idx;
-              int64_t in_idx = window_start + t;
-              out_data[out_idx * out_stride] = window_fn.data[t] * in_data[in_idx * in_stride];
+        for (int window_idx = 0; window_idx < nwindows_; window_idx++) {
+          for (int t = 0; t < window_length_; t++) {
+            int out_idx = window_idx * window_length_ + t;
+            int in_idx = window_idx * window_step_ + t;
+            if (in_idx < in_size) {
+              out_data[out_idx * out_stride] = in_data[in_idx * in_stride];
+            } else {
+              out_data[out_idx * out_stride] = 0;
             }
           }
         }
@@ -146,7 +139,11 @@ void ExtractWindowsCpu<OutputType, InputType, Dims>::Run(
 }
 
 template class ExtractWindowsCpu<float, float, 1>;  // 1-channel
+template class ExtractWindowsCpu<uint8_t, uint8_t, 1>;  // 1-channel
+
 template class ExtractWindowsCpu<float, float, 2>;  // n-channel
+template class ExtractWindowsCpu<uint8_t, uint8_t, 2>;  // n-channel
+
 
 }  // namespace window
 }  // namespace signal
