+++ dali/kernels/signal/window/extract_windows_cpu.h
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef DALI_KERNELS_SIGNAL_WINDOW_EXTRACT_WINDOWS_CPU_H_
-#define DALI_KERNELS_SIGNAL_WINDOW_EXTRACT_WINDOWS_CPU_H_
+#ifndef DALI_KERNELS_SIGNAL_WINDOW_EXTRACT_WINDOWS_H_
+#define DALI_KERNELS_SIGNAL_WINDOW_EXTRACT_WINDOWS_H_
 
 #include <memory>
 #include <complex>
@@ -31,16 +31,14 @@ namespace window {
 struct ExtractWindowsArgs {
   int64_t window_length = -1;
   int64_t window_step = -1;
-  int64_t window_center = -1;
-  int     axis = -1;
-  bool    reflect_pad = false;
+  int64_t in_time_axis = -1;
+  int64_t out_frame_axis = -1;
 
   inline bool operator==(const ExtractWindowsArgs& oth) const {
     return window_length == oth.window_length &&
            window_step == oth.window_step &&
-           window_center == oth.window_center &&
-           axis == oth.axis &&
-           reflect_pad == oth.reflect_pad;
+           in_time_axis == oth.in_time_axis &&
+           out_frame_axis == oth.out_frame_axis;
   }
 
   inline bool operator!=(const ExtractWindowsArgs& oth) const {
@@ -55,25 +53,6 @@ struct ExtractWindowsArgs {
  *
  * @param args.window_step Length of the step between windows. If not provided, win_length
  *        will be choosen (no overlap)
- *
- * @param args.axis Index of the axis representing the temporal dimension. If not provided,
- *        the last dimension will be used
- *
- * @param args.window_center If set, window centers will be adjusted with this offset.
- *        Note: window_center must be a value within 0 and window_length
- *        If not specified, window_center = window_length / 2 will be assumed
- *        With window_center = window_length / 2, window centers are placed at multiples of
- *          `window_step`.
- *        With window_center = 0, windows are aligned to the left (window start at multiples of
- *          `window_step`).
- *        With window_center = window_length, windows are aligned to the right (window start at
- *          multiples of `window_step`).
- *
- * @param args.reflect_pad Determines the padding policy when sampling out of bounds.
- *        If true, the signal will be padded with its own reflection.
- *        If false, the signal will be padded with zeros.
- *        This option is only relevant when `window_center` is greater than 0
- *
  */
 template <typename OutputType = float, typename InputType = float, int Dims = 1>
 class DLL_PUBLIC ExtractWindowsCpu {
@@ -88,23 +67,19 @@ class DLL_PUBLIC ExtractWindowsCpu {
 
   DLL_PUBLIC KernelRequirements Setup(KernelContext &context,
                                       const InTensorCPU<InputType, InputDims> &in,
-                                      const InTensorCPU<float, 1> &window_fn,
                                       const ExtractWindowsArgs &args);
 
   DLL_PUBLIC void Run(KernelContext &context,
                       const OutTensorCPU<OutputType, OutputDims> &out,
                       const InTensorCPU<InputType, InputDims> &in,
-                      const InTensorCPU<float, 1> &window_fn,
                       const ExtractWindowsArgs &args);
-
  private:
+  ExtractWindowsArgs args_;
   int64_t window_length_ = -1;
   int64_t window_step_ = -1;
-  int64_t window_fn_length_ = -1;
   int64_t nwindows_ = -1;
-  int axis_ = -1;
-  int64_t window_center_offset_ = 0;
-  bool    reflect_pad_ = false;
+  int64_t in_time_axis_ = -1;
+  int64_t out_frame_axis_ = -1;
 };
 
 }  // namespace window
@@ -112,4 +87,4 @@ class DLL_PUBLIC ExtractWindowsCpu {
 }  // namespace kernels
 }  // namespace dali
 
-#endif  // DALI_KERNELS_SIGNAL_WINDOW_EXTRACT_WINDOWS_CPU_H_
+#endif  // DALI_KERNELS_SIGNAL_WINDOW_EXTRACT_WINDOWS_H_
