+++ dali/kernels/signal/window/extract_windows_cpu_test.cc
@@ -18,8 +18,8 @@
 #include <complex>
 #include <cmath>
 #include "dali/kernels/scratch.h"
+#include "dali/kernels/signal/signal_kernel_utils.h"
 #include "dali/kernels/signal/window/extract_windows_cpu.h"
-#include "dali/kernels/common/utils.h"
 #include "dali/test/test_tensors.h"
 #include "dali/test/tensor_test_utils.h"
 
@@ -30,15 +30,14 @@ namespace window {
 namespace test {
 
 class ExtractWindowsCpuTest : public::testing::TestWithParam<
-  std::tuple<std::array<int64_t, 2>, int64_t, int64_t, int64_t, int64_t, bool>> {
+  std::tuple<std::array<int64_t, 2>, int64_t, int64_t, int64_t, int64_t>> {
  public:
   ExtractWindowsCpuTest()
     : data_shape_(std::get<0>(GetParam()))
     , window_length_(std::get<1>(GetParam()))
     , window_step_(std::get<2>(GetParam()))
-    , axis_(std::get<3>(GetParam()))
-    , window_center_(std::get<4>(GetParam()))
-    , reflect_pad_(std::get<5>(GetParam()))
+    , in_time_axis_(std::get<3>(GetParam()))
+    , out_frame_axis_(std::get<4>(GetParam()))
     , data_(volume(data_shape_))
     , in_view_(data_.data(), data_shape_) {}
 
@@ -46,44 +45,14 @@ class ExtractWindowsCpuTest : public::testing::TestWithParam<
 
  protected:
   void SetUp() final {
-    SequentialFill(in_view_, 0);
+    std::mt19937_64 rng;
+    UniformRandomFill(in_view_, rng, 0., 1.);
   }
   TensorShape<2> data_shape_;
-  int64_t window_length_ = -1, window_step_ = -1;
-  int axis_ = -1;
-  int64_t window_center_ = -1;
-  bool reflect_pad_ = false;
+  int64_t window_length_ = -1, window_step_ = -1, in_time_axis_ = -1, out_frame_axis_ = -1;
   std::vector<float> data_;
   OutTensorCPU<float, 2> in_view_;
 };
-
-template <typename T>
-void print_data(const OutTensorCPU<T, 2>& data_view) {
-  auto sh = data_view.shape;
-  for (int i0 = 0; i0 < sh[0]; i0++) {
-    for (int i1 = 0; i1 < sh[1]; i1++) {
-      int k = i0 * sh[1] + i1;
-      LOG_LINE << " " << data_view.data[k];
-    }
-    LOG_LINE << "\n";
-  }
-}
-
-template <typename T>
-void print_data(const OutTensorCPU<T, 3>& data_view) {
-  auto sh = data_view.shape;
-  for (int i0 = 0; i0 < sh[0]; i0++) {
-    for (int i1 = 0; i1 < sh[1]; i1++) {
-      for (int i2 = 0; i2 < sh[2]; i2++) {
-        int k = i0 * sh[1] * sh[2] + i1 * sh[2] + i2;
-        LOG_LINE << " " << data_view.data[k];
-      }
-      LOG_LINE << "\n";
-    }
-    LOG_LINE << "\n";
-  }
-  LOG_LINE << "\n";
-}
 
 TEST_P(ExtractWindowsCpuTest, ExtractWindowsTest) {
   using OutputType = float;
@@ -99,26 +68,16 @@ TEST_P(ExtractWindowsCpuTest, ExtractWindowsTest) {
   ExtractWindowsArgs args;
   args.window_length = window_length_;
   args.window_step = window_step_;
-  args.axis = axis_;
-  args.window_center = window_center_;
-  args.reflect_pad = reflect_pad_;
+  args.in_time_axis = in_time_axis_;
+  args.out_frame_axis = out_frame_axis_;
 
-  // Hamming window
-  std::vector<float> window_fn_data(window_length_);
-  for (int t = 0; t < window_length_; t++) {
-    window_fn_data[t] = 0.54f - 0.46f * cos(2.0f * M_PI * (t+0.5f) / window_length_);
-  }
-  auto window_fn_view = OutTensorCPU<float, 1>(window_fn_data.data(), {1});
+  KernelRequirements reqs = kernel.Setup(ctx, in_view_, args);
+  auto shape = reqs.output_shapes[0];
 
-  KernelRequirements reqs = kernel.Setup(ctx, in_view_, window_fn_view, args);
+  auto n = in_view_.shape[in_time_axis_];
+  auto nwindows = (n + window_step_ - 1) / window_step_;
+  auto expected_out_size = nwindows * window_length_ * volume(in_view_.shape) / n;
   auto out_shape = reqs.output_shapes[0][0];
-
-  auto n = in_view_.shape[axis_];
-  auto nwindows = n / window_step_ + 1;
-  auto expected_out_shape = TensorShape<DynamicDimensions>{
-    in_view_.shape[0], window_length_, nwindows};
-  ASSERT_EQ(expected_out_shape, out_shape);
-  auto expected_out_size = volume(expected_out_shape);
   auto out_size = volume(out_shape);
   ASSERT_EQ(expected_out_size, out_size);
 
@@ -129,61 +88,71 @@ TEST_P(ExtractWindowsCpuTest, ExtractWindowsTest) {
   TensorShape<> in_shape = in_view_.shape;
   TensorShape<> in_strides = GetStrides(in_shape);
 
-  TensorShape<> flat_out_shape = in_shape;
-  flat_out_shape[InputDims-1] = nwindows * window_length_;
-  TensorShape<> out_strides = GetStrides(flat_out_shape);
+  TensorShape<> out_tmp_shape = in_shape;
+  out_tmp_shape[InputDims-1] = nwindows * window_length_;
+  TensorShape<> out_strides = GetStrides(out_tmp_shape);
 
-  auto in_stride = in_strides[axis_];
-  auto out_stride = out_strides[axis_];
+  auto in_stride = in_strides[in_time_axis_];
+  auto out_stride = out_strides[in_time_axis_];
 
-  int64_t window_center_offset = window_center_ < 0 ? window_length_ / 2 : window_center_;
   for (int i = 0; i < in_view_.shape[0]; i++) {
     auto *out_slice = expected_out_view.data + i * out_strides[0];
     auto *in_slice = in_view_.data + i * in_strides[0];
     for (int w = 0; w < nwindows; w++) {
       for (int t = 0; t < window_length_; t++) {
-        auto out_k = w + t * nwindows;
-        auto in_k = w * window_step_ + t - window_center_offset;
-        if (reflect_pad_) {
-          while (in_k < 0 || in_k >= n) {
-              in_k = (in_k < 0) ? -in_k : 2*n-2-in_k;
-          }
-        }
-        out_slice[out_k] = (in_k >= 0 && in_k < n) ?
-          window_fn_data[t] * in_slice[in_k] : 0;
+        auto out_k = w * window_length_ + t;
+        auto in_k = w * window_step_ + t;
+        out_slice[out_k] = (in_k < n) ? in_slice[in_k] : 0;
       }
     }
   }
 
-
   LOG_LINE << "in:\n";
-  print_data(in_view_);
+  for (int i0 = 0; i0 < in_view_.shape[0]; i0++) {
+    for (int i1 = 0; i1 < in_view_.shape[1]; i1++) {
+      int k = i0*in_shape[1] + i1;
+      LOG_LINE << " " << in_view_.data[k];
+    }
+    LOG_LINE << "\n";
+  }
+  LOG_LINE << "\n";
 
-  LOG_LINE << "expected out:\n";
-  print_data(expected_out_view);
+
+  LOG_LINE << "out:\n";
+  auto sh = expected_out_view.shape;
+  for (int i0 = 0; i0 < sh[0]; i0++) {
+    for (int i1 = 0; i1 < sh[1]; i1++) {
+      for (int i2 = 0; i2 < sh[2]; i2++) {
+        int k = i0 * sh[1] * sh[2] + i1 * sh[2] + i2;
+        LOG_LINE << " " << expected_out_view.data[k];
+      }
+      LOG_LINE << "\n";
+    }
+    LOG_LINE << "\n";
+  }
+  LOG_LINE << "\n";
+
 
   std::vector<OutputType> out(out_size);
   auto out_view = OutTensorCPU<OutputType, OutputDims>(
     out.data(), out_shape.to_static<OutputDims>());
-  kernel.Run(ctx, out_view, in_view_, window_fn_view, args);
-
-  LOG_LINE << "out:\n";
-  print_data(out_view);
+  kernel.Run(ctx, out_view, in_view_, args);
 
   for (int idx = 0; idx < volume(out_view.shape); idx++) {
-    ASSERT_EQ(expected_out[idx], out_view.data[idx]) <<
-      "Output data doesn't match reference (idx=" << idx << ")";
+    EXPECT_EQ(expected_out[idx], out_view.data[idx]);
   }
 }
 
 INSTANTIATE_TEST_SUITE_P(ExtractWindowsCpuTest, ExtractWindowsCpuTest, testing::Combine(
-    testing::Values(std::array<int64_t, 2>{1, 12},
-                    std::array<int64_t, 2>{2, 12}),
-    testing::Values(4),  // window_length
-    testing::Values(2),  // step
-    testing::Values(1),  // axis
-    testing::Values(0, 2, 4),  // window offsets
-    testing::Values(true, false)));  // reflect padding
+    testing::Values(std::array<int64_t, 2>{3, 4}),
+                    //std::array<int64_t, 2>{1, 10}),
+                    //std::array<int64_t, 2>{1, 64},
+                    //std::array<int64_t, 2>{1, 100},
+                    //std::array<int64_t, 2>{1, 4096}),
+    testing::Values(4, 2),
+    testing::Values(2, 4),
+    testing::Values(1),
+    testing::Values(2)));
 
 }  // namespace test
 }  // namespace window
